requirements/production.txt
# Pro-tip: Try not to put anything here. 
# Avoid dependencies in production that aren't in development.
-r base.txt

# WSGI Handler
# ------------------------------------------------
gevent==1.2.1
gunicorn==19.7.1

# Static and Media Storage
# ------------------------------------------------
boto==2.46.1
django-storages-redux==1.3.2

# Email backends for Mailgun, Postmark, SendGrid and more
# -------------------------------------------------------
django-anymail==0.9

# Raven is the Sentry client
# --------------------------
raven==6.0.0
